addition of user authentication for some endpoint and permission checks

+++ app/main.py
@@ -5,9 +5,8 @@ import models
 import schemas
 import database
 from fastapi.security import OAuth2PasswordRequestForm
-from security import verify_password, create_access_token
+from security import verify_password, create_access_token, get_password_hash, get_current_active_user
 from sqlalchemy.exc import IntegrityError
-from security import get_password_hash
 
 # Create the database tables
 models.Base.metadata.create_all(bind=database.engine)
@@ -18,12 +17,16 @@ app = FastAPI(
     version="1.0.0",
     openapi_tags=[
         {
+            "name": "auth",
+            "description": "Operations for authentication. The **register** and **token** endpoints allow you to register and login.",
+        },
+        {
             "name": "ticket",
-            "description": "Operations with a single ticket. The **ticket** endpoint allows you to create, read, update and delete a ticket.",
+            "description": "Operations for a single ticket. The **ticket** endpoint allows you to create, read, update and delete a ticket.",
         },
         {
             "name": "tickets",
-            "description": "Operations with multiple tickets. The **tickets** endpoint allows you to read all tickets.",
+            "description": "Operations for multiple tickets. The **tickets** endpoint allows you to read all tickets.",
         }
     ]
 )
@@ -48,7 +51,7 @@ def register(user: schemas.UserCreate, db: Session = Depends(database.get_db)):
     return {"access_token": access_token, "token_type": "bearer"}
 
 
-@app.post("/login", tags=["auth"], response_model=schemas.Token)
+@app.post("/token", tags=["auth"], response_model=schemas.Token)
 def login(form_data: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(database.get_db)):
     client_id = db.query(models.User).filter(
         models.User.email == form_data.username).first().id
@@ -61,8 +64,9 @@ def login(form_data: OAuth2PasswordRequestForm = Depends(), db: Session = Depend
 
 
 @app.post("/ticket", tags=["ticket"], response_model=schemas.TicketPublic)
-def create_ticket(ticket: schemas.TicketCreate, db: Session = Depends(database.get_db)):
+def create_ticket(ticket: schemas.TicketCreate, current_user: models.User = Depends(get_current_active_user), db: Session = Depends(database.get_db)):
     db_ticket = models.Ticket(**ticket.model_dump())
+    db_ticket.author = current_user.id
     db.add(db_ticket)
     db.commit()
     db.refresh(db_ticket)
@@ -78,10 +82,13 @@ def read_ticket(ticket_id: UUID, db: Session = Depends(database.get_db)):
 
 
 @app.put("/ticket/{ticket_id}", tags=["ticket"], response_model=schemas.TicketPublic)
-def update_ticket(ticket_id: UUID, ticket: schemas.TicketUpdate, db: Session = Depends(database.get_db)):
+def update_ticket(ticket_id: UUID, ticket: schemas.TicketUpdate, current_user: models.User = Depends(get_current_active_user), db: Session = Depends(database.get_db)):
     db_ticket = db.get(models.Ticket, ticket_id)
     if not db_ticket:
         raise HTTPException(status_code=404, detail="Ticket not found")
+
+    if not current_user.role == schemas.Role.admin:
+        raise HTTPException(status_code=403, detail="Not enough permissions")
 
     update_data = ticket.model_dump(exclude_unset=True)
     for field, value in update_data.items():
@@ -93,47 +100,44 @@ def update_ticket(ticket_id: UUID, ticket: schemas.TicketUpdate, db: Session = D
 
 
 @app.delete("/ticket/{ticket_id}", tags=["ticket"])
-def delete_ticket(ticket_id: UUID, db: Session = Depends(database.get_db)):
+def delete_ticket(ticket_id: UUID, current_user: models.User = Depends(get_current_active_user), db: Session = Depends(database.get_db)):
     db_ticket = db.get(models.Ticket, ticket_id)
     if not db_ticket:
         raise HTTPException(status_code=404, detail="Ticket not found")
+
+    if not current_user.role == schemas.Role.admin:
+        raise HTTPException(status_code=403, detail="Not enough permissions")
 
     db.delete(db_ticket)
     db.commit()
     return {"message": "Ticket deleted successfully"}
 
 
-@app.put("/ticket/assign", tags=["ticket"], response_model=schemas.TicketPublic)
-def assign_ticket(ticket_id: UUID, assigned_to: UUID, db: Session = Depends(database.get_db)):
+@app.put("/ticket/{ticket_id}/assign", tags=["ticket"], response_model=schemas.TicketPublic)
+def assign_ticket(ticket_id: UUID, assign_to: UUID, current_user: models.User = Depends(get_current_active_user), db: Session = Depends(database.get_db)):
     db_ticket = db.get(models.Ticket, ticket_id)
     if not db_ticket:
         raise HTTPException(status_code=404, detail="Ticket not found")
 
-    db_ticket.assigned_to = assigned_to
+    if not ((current_user.role == schemas.Role.admin) or (current_user.id == assign_to and db_ticket.assigned_to == None)):
+        raise HTTPException(status_code=403, detail="Not enough permissions")
+
+    db_ticket.assigned_to = assign_to
     db.commit()
     db.refresh(db_ticket)
     return db_ticket
 
 
-@app.put("/ticket/status", tags=["ticket"], response_model=schemas.TicketPublic)
-def update_ticket_status(ticket_id: UUID, status: schemas.Status, db: Session = Depends(database.get_db)):
+@app.put("/ticket/{ticket_id}/status", tags=["ticket"], response_model=schemas.TicketPublic)
+def update_ticket_status(ticket_id: UUID, status: schemas.Status, current_user: models.User = Depends(get_current_active_user), db: Session = Depends(database.get_db)):
     db_ticket = db.get(models.Ticket, ticket_id)
     if not db_ticket:
         raise HTTPException(status_code=404, detail="Ticket not found")
+
+    if not ((current_user.role == schemas.Role.admin) or (current_user.id == db_ticket.author and status == schemas.Status.closed) or (current_user.id == db_ticket.assigned_to)):
+        raise HTTPException(status_code=403, detail="Not enough permissions")
 
     db_ticket.status = status
-    db.commit()
-    db.refresh(db_ticket)
-    return db_ticket
-
-
-@app.put("/ticket/priority", tags=["ticket"], response_model=schemas.TicketPublic)
-def update_ticket_priority(ticket_id: UUID, priority: schemas.Priority, db: Session = Depends(database.get_db)):
-    db_ticket = db.get(models.Ticket, ticket_id)
-    if not db_ticket:
-        raise HTTPException(status_code=404, detail="Ticket not found")
-
-    db_ticket.priority = priority
     db.commit()
     db.refresh(db_ticket)
     return db_ticket
